+++ packages/ckeditor5-core/src/editor/editor.ts
@@ -11,7 +11,6 @@ import {
 	Config,
 	CKEditorError,
 	ObservableMixin,
-	logWarning,
 	type Locale,
 	type LocaleTranslate,
 	type ObservableChangeEvent
@@ -281,6 +280,17 @@ export default abstract class Editor extends /* #__PURE__ */ ObservableMixin() {
 	constructor( config: EditorConfig = {} ) {
 		super();
 
+		if ( 'sanitizeHtml' in config ) {
+			/**
+			 * Configuration property `config.sanitizeHtml` was removed in CKEditor version 43.1.0 and is no longer supported.
+			 *
+			 * Please use `config.htmlEmbed.sanitizeHtml` and/or `config.mergeFields.sanitizeHtml` instead.
+			 *
+			 * @error editor-config-sanitizehtml-not-supported
+			 */
+			throw new CKEditorError( 'editor-config-sanitizehtml-not-supported' );
+		}
+
 		const constructor = this.constructor as typeof Editor;
 
 		// We don't pass translations to the config, because its behavior of splitting keys
@@ -301,23 +311,6 @@ export default abstract class Editor extends /* #__PURE__ */ ObservableMixin() {
 		this.config = new Config<EditorConfig>( rest, defaultConfig );
 		this.config.define( 'plugins', availablePlugins );
 		this.config.define( this._context._getEditorConfig() );
-		this.config.define( 'sanitizeHtml', function( rawHtml ) {
-			/**
-			 * One of the editor features directly inserts unsanitized HTML code into the editor.
-			 * It is strongly recommended to define a sanitize function that will clean up the input HTML
-			 * in order to avoid XSS vulnerability.
-			 *
-			 * For a detailed overview, check the {@glink getting-started/setup/html-security "HTML security"} guide.
-			 *
-			 * @error provide-sanitize-function
-			 */
-			logWarning( 'provide-sanitize-function' );
-
-			return {
-				html: rawHtml,
-				hasChanged: false
-			};
-		} );
 
 		this.plugins = new PluginCollection<Editor>( this, availablePlugins, this._context.plugins );