+++ packages/ckeditor5-html-embed/tests/htmlembedediting.js
@@ -73,47 +73,41 @@ describe( 'HtmlEmbedEditing', () => {
 	} );
 
 	describe( 'config', () => {
-		let htmlEmbedConfig;
+		let htmlEmbed;
 
 		beforeEach( () => {
-			htmlEmbedConfig = editor.config.get( 'htmlEmbed' );
+			htmlEmbed = editor.config.get( 'htmlEmbed' );
 		} );
 
 		describe( 'htmlEmbed.showPreviews', () => {
 			it( 'should be set to `false` by default', () => {
-				expect( htmlEmbedConfig.showPreviews ).to.equal( false );
+				expect( htmlEmbed.showPreviews ).to.equal( false );
 			} );
 		} );
 
 		describe( 'htmlEmbed.sanitizeHtml', () => {
-			it( 'should override the default sanitizer from `config.sanitizeHtml` if provided and log the warning', async () => {
-				testUtils.createSinonSandbox();
-
-				const element = document.createElement( 'div' );
-				document.body.appendChild( element );
+			beforeEach( () => {
 				sinon.stub( console, 'warn' );
+			} );
 
-				const editor = await ClassicTestEditor
-					.create( element, {
-						plugins: [ HtmlEmbedEditing ],
-						htmlEmbed: {
-							showPreviews: true,
-							// The default sanitize function without `console.warn`.
-							sanitizeHtml: input => ( { html: input, hasChanged: false } )
-						}
-					} );
-				const configuredSanitizer = editor.config.get( 'htmlEmbed.sanitizeHtml' );
+			it( 'should return an object with cleaned html and a note whether something has changed', () => {
+				expect( htmlEmbed.sanitizeHtml( 'foo' ) ).to.deep.equal( {
+					html: 'foo',
+					hasChanged: false
+				} );
+			} );
 
-				configuredSanitizer( 'foo' );
+			it( 'should return an input string (without any modifications)', () => {
+				const unsafeHtml = '<img src="data:/xxx,<script>void</script>" onload="void;">';
 
-				// Only the warning about the deprecated option should be logged.
+				expect( htmlEmbed.sanitizeHtml( unsafeHtml ).html ).to.deep.equal( unsafeHtml );
+			} );
+
+			it( 'should display a warning when using the default sanitizer', () => {
+				htmlEmbed.sanitizeHtml( 'foo' );
+
 				expect( console.warn.callCount ).to.equal( 1 );
-				expect( console.warn.firstCall.args[ 0 ] ).to.equal(
-					'The `htmlEmbed.sanitizeHtml` configuration option is deprecated. Use the `sanitizeHtml` option instead.'
-				);
-
-				await editor.destroy();
-				element.remove();
+				expect( console.warn.firstCall.args[ 0 ] ).to.equal( 'html-embed-provide-sanitize-function' );
 			} );
 		} );
 	} );
@@ -697,9 +691,9 @@ describe( 'HtmlEmbedEditing', () => {
 					.create( element, {
 						plugins: [ HtmlEmbedEditing ],
 						htmlEmbed: {
-							showPreviews: true
-						},
-						sanitizeHtml
+							showPreviews: true,
+							sanitizeHtml
+						}
 					} )
 					.then( newEditor => {
 						editor = newEditor;