+++ tests/manual/all-features.js
@@ -167,9 +167,9 @@ ClassicEditor
 			}
 		},
 		htmlEmbed: {
-			showPreviews: true
+			showPreviews: true,
+			sanitizeHtml: html => ( { html, hasChange: false } )
 		},
-		sanitizeHtml: html => ( { html, hasChange: false } ),
 		list: {
 			properties: {
 				styles: true,